requirement.txt
ray[all]
webdataset
librosa
numpy
tqdm
pydub
pandas
descript-audiotools
huggingface_hub[hf_transfer]
gpustat